Compile README.md via markdown-unlit
Keeping a `README.md` up to date with your library is an annoying
endeavor. They are constantly falling out of sync. Luckily we have type
checking to ammend this for us. By utilizing `markdown-unlit` we can
compile the `README.md` for this project and keep it in sync with
library updates. This keeps our examples "working" and ensures less
dissonance for users.

This toolchain involves actually storing `README.md` details in
`README.lhs` and then symlinking `README.md` to that file. `github` and
other tools pick up on the symlink and transparently display the
appropriate markdown.