Throw exceptions from heartbeat to parent thread
The heartbeat process can fail. This causes its individual thread to
stop, but does not crash the entire process. The threading behavior of
heartbeat should be transparent to users of the library, so an exception
there should throw an exception to the host.

This specifically helps prevent issues where a lost socket is caught in
silenced in the heartbeat and prevents the entire client from crashing.