Supply Queue in JobOptions, not separately
Also adds support for Jobtype in JobOptions and changes the default
value from "Example" to "Default".

NOTE: "Job type" is formatted as Jobtype in the Faktory protocol, so we
respect that here with Jobtype/jobtype, even though JobType/jobType
would probably be better. We also use terse local variables q and jt to
avoid clashing with the setter names.