Separate no-job from invalid-job cases
Upgrade commandJSON to Either String (Maybe a) so JSON parsing failures
can be captured in Left. We still need Maybe within the Right branch
since no Job at all is a normal part of consuming and Nothing is a
fitting representation.

Then pass that up through fetchJob and handle the Left case specifically
in processorLoop. We still ignore can carry on, but now with a log
message. In an ideal world, we would attempt to re-parse the Job for
just its Id, so we could Fail it formally rather than leave it to time
out -- but that's very hard to do with the way the functions are broken
up right now, so has been deferred.