Move ghc-options to package.yaml
In a megarepo, it makes sense to put options in the one stack.yaml to be
used by the many package.yamls. That is where I originally copied this
pattern from. However, in a repo with one package.yaml and a CI matrix
of many stack.yamls, it makes more sense to do the reverse.