Stop hiding Left results in recvUnsafe
Previously, when this function was unable to parse a Reply from Faktory,
it would log it and return a Nothing. This takes what should be a
serious error and replaces it by something innocuous-looking or even
expected.

In general, recvUnsafe shouldn't make the assumption that its callers
don't care about the Lefts and are fine with the log+Nothing handling
it's been taking internally. Concretely, the ramifications depend on
use-case:

In command_, the return value from recvUnsafe is completely discarded,
things will appear successful when they very much were not. That is not
changed by this commit, but maintaining the information of the error
beyond recvUnsafe is required to fix it next.

In commandOK, we want to include failure information in the error we
throw. Again, this requires recvUnsafe not destroy that information
internally.

In commandJSON as used by fetchJob, a Nothing represents no Job ready
and that the worker should idle and loop. Masking failures as no Job
here is mostly fine, but handling the Left with log+ignore here, vs
internally in recvUnsafe, is at least better in principle.

It's worth noting that the "unsafe" in recvUnsafe's naming is referring
to thread-safety, and not anything to do with handling failures, or any
lack thereof.