Add some explicit bounds
Using of --pvp-bounds will add them in accordance with the version the
in-use snapshot. Depending on a variety of circumstances this can be too
restrictive. In this case these dependencies need relaxing to build in
Nightly and remain in current Stackage. Adding manual bounds (only when
required like this) is the recommended approach.