Move delays in batch tests
These delays, in their current location, are actually before the BATCH
COMMIT. That means they don't do anything to help give Faktory time to
handle the callbacks we want to test.

The two-callback case was just slow enough to reliably lose the race,
making me think that was just not working. However, moving the delay
outside of the batch makes it all work, as intended.

I also reduced the delay, since things seem to be reliably green at 0.5
seconds, but not at 0.25s.