Fix parsing of "at"
When we moved to handling at-vs-in_ to an Either representation, the
JSON parser for reading back the at-value as a resolved UTCTime was
incorrectly implemented to construct with Right instead of Left.

This meant expecting a NominimalDiffTime instead, which fails.

    [ERROR]: Invalid Job: Error in $.at: parsing NominalDiffTime failed,
    expected Number, but encountered String

Jobs not enqueued with at (or in_) worked fine, which is how our tests
missed it.