Stop releasing with restrictive upper bounds
This has been nothing but a maintenance burden. Whenever they actually
restrict something, we reflexively update them. Might well have that
just be the behavior.

This leaves a small gap in that we may not compile with a newer
dependency that we claim to. This gap existed to some degree always, and
we rely mostly on the Stackage build processes to notify us to it. This
change increases the gap, which we are comfortable with, but we do hope
to enhance CI in the future to explicitly test for it.