Deprecate trackPerform
Defining this function was unnecessary. Providing a JobOptions helper
fits much better with the overall library as is not too much extra
characters:

    trackPerform options producer job
    ==
    perform (options <> tracked) producer job

Exporting trackPerform actually fit poorly in our own usage, nudging us
the wrong way and ending up with tracked Jobs not going through our
centralized enqueue function (which does things like defaulting
options).

Had we exposed tracked to begin with, it would've been obvious to use
our centralized enqueue but just append the tracked option as needed.

At this commit, the tests are still using it, in order to display the
deprecation for review:

    .../tests/Faktory/Ent/TrackingSpec.hs:43:14: error: [-Wdeprecations, -Werror=deprecations]
        In the use of ‘trackPerform’ (imported from Faktory.Ent.Tracking):
        Deprecated: "Use ‘perform (options <> tracked)’ instead"
       |
    43 |         a <- trackPerform @Text mempty producer "a"
       |              ^^^^^^^^^^^^