Add jobRetriesRemaining
Ultimately, we want this to reduce alerting noise by only notifying to
Bugsnag of Job exceptions once jobRetriesRemaining == 0.

I chose to parse the whole JobFailure value since we are a library and
others may find those bits useful. Also, it was a trivial parse with
mostly nullable fields.

I wrote a temporary spec that enqueued a Job with retries and verified
lots of actual behavior as they were processed, but I am not committing
it because it takes forever to run, since it has to actually wait for
Faktory's retry delays.

See https://github.com/contribsys/faktory/issues/374 for more details.