Prevent zombie consumers by respecting reserve_for
`reserve_for` is an optional configuration for jobs. If it is not
provided the default is 1800 seconds. This value describes the window in
which a consumer may process and job and send back an `ACK`. If an `ACK`
is not sent in this period then Faktory will hand the job to another
consumer.

Not respecting this TTL can result in __zombie__ jobs. A worker that
exceeds the `reserve_for` will continue functioning while Faktory may
have passed the job to another worker. This can cause duplicate effort,
performance issues, and logical bugs.

This change simply places a `timeout` around the job processing time.
This ensures the job is appropriately timed. The current implementation
is silent, no information is logged, the job is just terminated. That is
likely not desirable.