Allow configuration of default retry and queue
Faktory has defaults for `queue: default` and `retry: 25`. This means
that `queue` and `retry` are not required settings, however these
defaults may not be desired, but the optionality makes it painful to
enforce across a project. This enhancement allows these defaults values
to be configured within `Settings` data. This allows `queue` and `retry`
to remain optional, but fall back to a project defined default.