Allow configuration of default job options
Faktory has defaults for `queue: default` and `retry: 25`. This means
that `queue` and `retry` are not required settings, but these defaults
may not be desired by the end user. Unfortunately the optionality makes
it painful to enforce across a project. This enhancement allows these
and other defaults job options to be configured within `Settings` data.
This allows `queue` and `retry` to remain optional, but fall back to a
project defined default when prescribed.