Add HasJobType to correctly report `jobType` to Faktory
`jobType` within Faktory is not heavily utilized by this library,
however it is useful from a reporting perspective. This enhancement adds
a `HasJobType` class that ensures any `ToJSON` passed to `perform` also
implements a `jobTypeName` to be reported in Faktory. This is an easy
constraint to satisfy, but it does cause a breaking change.