Use Typeable to default jobtype name
The `jobtype` option in faktory is not heavily utilized by this
implementation because sum types and their interplay with `aeson`
provide an easy facility to differentiate jobs. However `jobtype` is
useful from a logging and reporting perspective. This enhancement
utilizes `Typeable` a default derived typeclass in Haskell to pull the
type constructor name of the job datatype as a default. This default can
always be overriden by using `jobtype` in job options, but falls back to
this usable default.