Changes FaktoryException's Show to tag that it's a FaktoryException
Previously this type was using `newtype` deriving which just passes through the calls and so the `Show` instance was just printing the internals which lead me to mistake it for a StringException itself.

This will use stock `Show` which will be `show (FaktoryException e) = "FaktoryException " <> show e`.  That gives it a nice little prefix.