Fail Jobs that don't decode as expected
Originally, this decoding failure would enter the `Left` case and only
log the error. The Job would hang, neither `ACK`ed nor `FAIL`ed until
its reservation expired.

Now, as long as it is valid JSON, we reach the `Right`-`Just` branch and
attempt to decode it further there. This way, if it can't decode, that
exception occurs in the right place to `FAIL` the job, as desired.

Closes #88.