Use Default TLSSettings, support crypton-connection >= 0.4.0
We were re-specifying the defaults already[^1], which was fine until
0.4.0 added a new member of a type that is not easily constructed
(`Supported`). In fact, the library seems to want you to construct it
via the `Default` instance (which we don't normally prefer), so we might
as well do the whole record like that to avoid any CPP on the specific
field.

[^1]: https://hackage.haskell.org/package/crypton-connection-0.4.3/docs/src/Network.Connection.Types.html#line-95