Support CI without Faktory credentials
Adds a `prep` job that checks to see if the `secrets.FAKTORY` values are
available and, only if so, uses the enterprise image and runs the
`Faktory.Ent` specs. Otherwise, if we use a free image and skip those
tests, then CI works for Dependabot and fork PRs, for example.

We could've solved Dependabot by making the necessary secrets available
in this scope, but it seems good to solve for forks as well.